Refactor Example project
