+++ Example/Artisan/Common/EventRouter.swift
