+++ Example/Artisan/Details/Component/EventTitleView.swift
