+++ Example/Artisan/Details/Component/SimilarEventView.swift
@@ -14,6 +14,8 @@ import Pharos
 import Builder
 import Impose
 
+// MARK: ViewModel Protocol
+
 protocol SimilarEventViewDataBinding {
     var eventsObservable: Observable<[SimilarEvent]> { get }
 }
@@ -22,12 +24,16 @@ protocol SimilarEventViewSubscriber {
     func didTap(_ event: SimilarEvent, at indexPath: IndexPath)
 }
 
+typealias SimilarEventViewModel = ViewModel & SimilarEventViewDataBinding & SimilarEventViewSubscriber
+
+// MARK: Intermediate Model
+
 struct SimilarEvent: IntermediateModel {
     var distinctifier: AnyHashable
     var viewModel: EventCollectionCellViewModel
 }
 
-typealias SimilarEventViewModel = ViewModel & SimilarEventViewDataBinding & SimilarEventViewSubscriber
+// MARK: View
 
 class SimilarEventView: UIPlannedView, ViewBindable {
     typealias Model = SimilarEventViewModel
@@ -86,49 +92,5 @@ extension SimilarEventView: UICollectionViewDelegate {
         collectionView.deselectItem(at: indexPath, animated: true)
         guard indexPath.item < events.count else { return }
         model?.didTap(events[indexPath.item], at: indexPath)
-    }
-}
-
-class SimilarEventVM: SimilarEventViewModel, ObjectRetainer {
-    typealias TapAction = (SimilarEventVM, Event) -> Void
-    
-    @Injected var service: EventService
-    
-    @Subject var event: Event?
-    @Subject var events: [Event] = []
-    var eventsObservable: Observable<[SimilarEvent]> {
-        $events.mapped { events in
-            events.compactMap {
-                SimilarEvent(
-                    distinctifier: $0,
-                    viewModel: EventCollectionCellVM(event: $0)
-                )
-            }
-        }
-        
-    }
-    
-    private var tapObserver: TapAction?
-    
-    init(event: Event?) {
-        self.event = event
-        $event.whenDidSet { [unowned self] changes in
-            guard let new = changes.new else { return }
-            service.similarEvent(with: new) { [weak self] similars in
-                self?.events = similars
-            }
-        }.retained(by: self)
-            .fire()
-    }
-    
-    @discardableResult
-    func whenDidTapped(thenRun action: @escaping TapAction) -> Self {
-        tapObserver = action
-        return self
-    }
-    
-    func didTap(_ event: SimilarEvent, at indexPath: IndexPath) {
-        guard let tappedEvent = event.distinctifier as? Event else { return }
-        tapObserver?(self, tappedEvent)
     }
 }
