+++ Example/Artisan/Model/Event.swift
@@ -1,0 +1,17 @@
+//
+//  Event.swift
+//  Artisan_Example
+//
+//  Created by Nayanda Haberty on 06/06/22.
+//  Copyright © 2022 CocoaPods. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+struct Event: Hashable {
+    var image: UIImage
+    var name: String
+    var details: String
+    var date: Date
+}
